refactor(todel): isolate session types into their own file
