+++ todel/src/models/mod.rs
@@ -5,6 +5,7 @@ mod gateway;
 mod info;
 mod messages;
 mod response;
+mod sessions;
 mod users;
 
 pub use files::*;
@@ -12,6 +13,7 @@ pub use gateway::*;
 pub use info::*;
 pub use messages::*;
 pub use response::*;
+pub use sessions::*;
 pub use users::*;
 
 #[cfg(feature = "logic")]
